+++ test/unit/fuzz-tools.test.js
@@ -412,6 +412,7 @@ import * as badenWalkerV432 from '../../lib/baden-walker-v432.js';
 import * as modicChangesV433 from '../../lib/modic-changes-v433.js';
 import * as pfirrmannDiscV434 from '../../lib/pfirrmann-disc-v434.js';
 import * as vanHerickV435 from '../../lib/van-herick-v435.js';
+import * as bifflBcviV436 from '../../lib/biffl-bcvi-v436.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -769,6 +770,7 @@ const MODULES = {
   'modic-changes-v433.js': modicChangesV433,
   'pfirrmann-disc-v434.js': pfirrmannDiscV434,
   'van-herick-v435.js': vanHerickV435,
+  'biffl-bcvi-v436.js': bifflBcviV436,
   'tb-testing.js': tbTesting,
 };